fixed autoretry for class based tasks
