+++ celery/app/autoretry.py
@@ -1,0 +1,50 @@
+from vine.utils import wraps
+
+from celery.exceptions import Ignore, Retry
+from celery.utils.time import get_exponential_backoff_interval
+
+
+def add_autoretry_behaviour(task, **options):
+    autoretry_for = tuple(
+        options.get('autoretry_for',
+                    getattr(task, 'autoretry_for', ()))
+    )
+    retry_kwargs = options.get(
+        'retry_kwargs', getattr(task, 'retry_kwargs', {})
+    )
+    retry_backoff = int(
+        options.get('retry_backoff',
+                    getattr(task, 'retry_backoff', False))
+    )
+    retry_backoff_max = int(
+        options.get('retry_backoff_max',
+                    getattr(task, 'retry_backoff_max', 600))
+    )
+    retry_jitter = options.get(
+        'retry_jitter', getattr(task, 'retry_jitter', True)
+    )
+
+    if autoretry_for and not hasattr(task, '_orig_run'):
+
+        @wraps(task.run)
+        def run(*args, **kwargs):
+            print('running ', task.name)
+            try:
+                return task._orig_run(*args, **kwargs)
+            except Ignore:
+                # If Ignore signal occures task shouldn't be retried,
+                # even if it suits autoretry_for list
+                raise
+            except Retry:
+                raise
+            except autoretry_for as exc:
+                if retry_backoff:
+                    retry_kwargs['countdown'] = \
+                        get_exponential_backoff_interval(
+                            factor=retry_backoff,
+                            retries=task.request.retries,
+                            maximum=retry_backoff_max,
+                            full_jitter=retry_jitter)
+                raise task.retry(exc=exc, **retry_kwargs)
+
+        task._orig_run, task.run = task.run, run
