+++ celery/app/base.py
@@ -16,14 +16,13 @@ from kombu.utils.compat import register_after_fork
 from kombu.utils.objects import cached_property
 from kombu.utils.uuid import uuid
 from vine import starpromise
-from vine.utils import wraps
 
 from celery import platforms, signals
 from celery._state import (_announce_app_finalized, _deregister_app,
                            _register_app, _set_current_app, _task_stack,
                            connect_on_app_finalize, get_current_app,
                            get_current_worker_task, set_default_app)
-from celery.exceptions import AlwaysEagerIgnored, ImproperlyConfigured, Ignore, Retry
+from celery.exceptions import AlwaysEagerIgnored, ImproperlyConfigured
 from celery.five import (UserDict, bytes_if_py2, python_2_unicode_compatible,
                          values)
 from celery.loaders import get_loader_cls
@@ -35,8 +34,7 @@ from celery.utils.functional import first, head_from_fun, maybe_list
 from celery.utils.imports import gen_task_name, instantiate, symbol_by_name
 from celery.utils.log import get_logger
 from celery.utils.objects import FallbackContext, mro_lookup
-from celery.utils.time import (get_exponential_backoff_interval, timezone,
-                               to_utc)
+from celery.utils.time import timezone, to_utc
 
 # Load all builtin tasks
 from . import builtins  # noqa
@@ -44,6 +42,7 @@ from . import backends
 from .annotations import prepare as prepare_annotations
 from .defaults import DEFAULT_SECURITY_DIGEST, find_deprecated_settings
 from .registry import TaskRegistry
+from .autoretry import add_autoretry_behaviour
 from .utils import (AppPickler, Settings, _new_key_to_old, _old_key_to_new,
                     _unpickle_app, _unpickle_app_v2, appstr, bugreport,
                     detect_settings)
@@ -462,49 +461,7 @@ class Celery(object):
                 pass
             self._tasks[task.name] = task
             task.bind(self)  # connects task to this app
-
-            autoretry_for = tuple(
-                options.get('autoretry_for',
-                            getattr(task, 'autoretry_for', ()))
-            )
-            retry_kwargs = options.get(
-                'retry_kwargs', getattr(task, 'retry_kwargs', {})
-            )
-            retry_backoff = int(
-                options.get('retry_backoff',
-                            getattr(task, 'retry_backoff', False))
-            )
-            retry_backoff_max = int(
-                options.get('retry_backoff_max',
-                            getattr(task, 'retry_backoff_max', 600))
-            )
-            retry_jitter = options.get(
-                'retry_jitter', getattr(task, 'retry_jitter', True)
-            )
-
-            if autoretry_for and not hasattr(task, '_orig_run'):
-
-                @wraps(task.run)
-                def run(*args, **kwargs):
-                    try:
-                        return task._orig_run(*args, **kwargs)
-                    except Ignore:
-                        # If Ignore signal occures task shouldn't be retried,
-                        # even if it suits autoretry_for list
-                        raise
-                    except Retry:
-                        raise
-                    except autoretry_for as exc:
-                        if retry_backoff:
-                            retry_kwargs['countdown'] = \
-                                get_exponential_backoff_interval(
-                                    factor=retry_backoff,
-                                    retries=task.request.retries,
-                                    maximum=retry_backoff_max,
-                                    full_jitter=retry_jitter)
-                        raise task.retry(exc=exc, **retry_kwargs)
-
-                task._orig_run, task.run = task.run, run
+            add_autoretry_behaviour(task, **options)
         else:
             task = self._tasks[name]
         return task
